extract UpdateConversationName
